+++ src/script/E2EIdentity/EnrollmentTimer/EnrollmentTimer.test.ts
@@ -17,7 +17,7 @@
  *
  */
 
-import {getEnrollmentTimer, messageRetentionTime} from './delay';
+import {getEnrollmentTimer, messageRetentionTime} from './EnrollmentTimer';
 
 import {MLSStatuses} from '../E2EIdentityVerification';
